+++ NEWS.md
@@ -1,3 +1,10 @@
+# ucrstats 0.2.0
+
+* Linear regression tutorials added
+
+* Linear regression vignette added
+
+
 # ucrstats 0.1.0
 
 * First deployment of the R package